Update app.py gael

+++ app.py
@@ -116,26 +116,67 @@ def valid_edit_etudiant():
 @app.route('/etudiant/show')
 def show_etudiants():
     mycursor = get_db().cursor()
-    sql=''' SELECT Etudiant.id_etudiant AS idEtudiant, Etudiant.nom AS nomEtudiant, Etudiant.prenom AS prenomEtudiant, Etudiant.email, Etudiant.telephone
-    FROM Etudiant
-    ORDER BY id_etudiant;'''
+    sql='''SELECT E.nom_etablissement, F.nom_formation, C.id_composante, Etudiant.id_etudiant AS idEtudiant, Etudiant.nom AS nomEtudiant, Etudiant.prenom AS prenomEtudiant, Etudiant.email AS emailEtudiant, Etudiant.telephone AS telephoneEtudiant
+FROM Etudiant
+JOIN Composante C on Etudiant.id_composante = C.id_composante
+JOIN Etablissement E on C.id_etablissement = E.id_etablissement
+JOIN Formation F on C.id_formation = F.id_formation
+ORDER BY idEtudiant;'''
     mycursor.execute(sql)
     liste_etudiants = mycursor.fetchall()
     return render_template('etudiant/show_etudiants.html', etudiants=liste_etudiants)
 
 @app.route('/etudiant/delete')
 def delete_etudiant():
-    id=request.args.get('id',0)
+    id = request.args.get('id', 0)
     print('''suppression du contrat avec l'ID : ''' + id)
     mycursor = get_db().cursor()
-    tuple_param=(id)
-    sql="""DELETE FROM Contrat WHERE id_etudiant=%s;"""
-    mycursor.execute(sql,tuple_param)
-    get_db().commit()
-    sql = """DELETE FROM Etudiant WHERE id_etudiant=%s;"""
+
+    # Use a tuple with a single element
+    tuple_param = (id,)
+
+    sql = """SELECT id_contrat, id_etudiant
+             FROM Contrat 
+             WHERE id_etudiant=%s;"""
     mycursor.execute(sql, tuple_param)
-    get_db().commit()
+    contrats = mycursor.fetchall()
+    print(contrats)
+
+    if len(contrats) == 0:
+        mycursor = get_db().cursor()
+
+        # Use a tuple with a single element
+        tuple_param = (id,)
+
+        sql = """DELETE FROM Etudiant WHERE id_etudiant=%s"""
+        mycursor.execute(sql, tuple_param)
+        get_db().commit()
+    else:
+        mycursor = get_db().cursor()
+
+        # Use a tuple with a single element
+        tuple_param = (id,)
+
+        sql = '''SELECT id_contrat, date_debut, date_fin, id_etudiant
+                 FROM Contrat
+                 WHERE id_etudiant = %s;'''
+        mycursor.execute(sql, tuple_param)
+        contrats = mycursor.fetchall()
+
+        sql = '''SELECT E.nom_etablissement, F.nom_formation, C.id_composante, Etudiant.id_etudiant AS idEtudiant, Etudiant.nom AS nomEtudiant, Etudiant.prenom AS prenomEtudiant, Etudiant.email AS emailEtudiant, Etudiant.telephone AS telephoneEtudiant
+                 FROM Etudiant
+                 JOIN Composante C on Etudiant.id_composante = C.id_composante
+                 JOIN Etablissement E on C.id_etablissement = E.id_etablissement
+                 JOIN Formation F on C.id_formation = F.id_formation
+                 WHERE Etudiant.id_etudiant = %s
+                 ORDER BY idEtudiant;'''
+
+        mycursor.execute(sql, tuple_param)
+        etudiants = mycursor.fetchall()
+        return render_template('etudiant/delete_etudiant.html', contrats=contrats, etudiants=etudiants)
+
     return redirect('/etudiant/show')
+
 
 @app.route('/etudiant/add', methods=['GET'])
 def add_etudiant():
@@ -145,25 +186,88 @@ def add_etudiant():
     ORDER BY id_etudiant;'''
     mycursor.execute(sql)
     liste_etudiants = mycursor.fetchall()
+    mycursor = get_db().cursor()
+    sql = '''SELECT nom_formation, id_formation
+       FROM Formation
+       ORDER BY id_formation;'''
+    mycursor.execute(sql)
+    composantes = mycursor.fetchall()
+    sql = """SELECT id_etablissement AS id_etablissement, nom_etablissement
+    FROM Etablissement"""
+    mycursor.execute(sql)
+    etablissements = mycursor.fetchall()
     print('''affichage du formulaire pour saisir un contrat''')
-    return render_template('contrat/add_contrat.html', etudiants=liste_etudiants)
+    return render_template('etudiant/add_etudiant.html', etudiants=liste_etudiants, composantes=composantes, etablissements=etablissements)
 
 @app.route('/etudiant/add', methods=['POST'])
 def valid_add_etudiant():
     print('''ajout du contrat dans le tableau''')
-    idVelo = request.form.get('idVelo')
     idEtudiant = request.form.get('idEtudiant')
-    dateDebut = request.form.get('dateDebut')
-    dateFin = request.form.get('dateFin')
+    nomEtudiant = request.form.get('nomEtudiant')
+    prenomEtudiant = request.form.get('prenomEtudiant')
+    emailEtudiant = request.form.get('emailEtudiant')
+    telephoneEtudiant = request.form.get('telephoneEtudiant')
+    idComposante = request.form.get('idComposante')
     #message = 'nom :' + nom + ' - groupe :' + groupe
     #print(message)
     mycursor = get_db().cursor()
-    tuple_param=(dateDebut, dateFin,idEtudiant, idVelo)
-    sql="INSERT INTO Etudiant (date_debut, date_fin, id_etudiant, id_velo) VALUES (%s, %s,%s,%s);"
+    tuple_param=(nomEtudiant, prenomEtudiant, emailEtudiant, telephoneEtudiant, idEtudiant, idComposante)
+    sql="INSERT INTO Etudiant (nom, prenom, email, telephone, id_etudiant, id_composante ) VALUES (%s,%s,%s,%s,%s,%s);"
     mycursor.execute(sql,tuple_param)
     get_db().commit()
-    return redirect('/contrat/show')
+    return redirect('/etudiant/show')
 
+@app.route('/etudiant/edit', methods=['GET'])
+def edit_etudiant():
+    print('''affichage du formulaire pour modifier un étudiant''')
+    print(request.args)
+    print(request.args.get('id'))
+    id = request.args.get('id')
+    mycursor = get_db().cursor()
+    sql = ''' SELECT id_etudiant AS id, nom, prenom, email, telephone, id_composante
+        FROM Etudiant
+        WHERE id_etudiant=%s;'''
+    tuple_param = (id)
+    mycursor.execute(sql, tuple_param)
+    etudiant = mycursor.fetchone()
+    sql = ''' SELECT Formation.id_formation AS id_formation, Formation.nom_formation AS nom_formation, C.id_etablissement AS id_etablissement, nom_etablissement
+    FROM Formation
+    JOIN Composante C on Formation.id_formation = C.id_formation
+    JOIN Etablissement E on E.id_etablissement = C.id_etablissement;'''
+    mycursor.execute(sql)
+    formations = mycursor.fetchall()
+    print(formations)
+    sql =  """SELECT id_etablissement AS id_etablissement, nom_etablissement
+    FROM Etablissement"""
+    mycursor.execute(sql)
+    etablissements = mycursor.fetchall()
+    return render_template('etudiant/edit_etudiant.html', etudiant=etudiant, formations=formations, etablissements=etablissements)
+
+@app.route('/etudiant/edit', methods=['POST'])
+def valid_edit_etudiant():
+    print('''modification de l'étudiant dans le tableau''')
+    idEtudiant = request.form.get('idEtudiant')
+    print(idEtudiant)
+    nomEtudiant = request.form.get('nomEtudiant')
+    print(nomEtudiant)
+    prenomEtudiant = request.form.get('prenomEtudiant')
+    print(prenomEtudiant)
+    emailEtudiant = request.form.get('emailEtudiant')
+    print(emailEtudiant)
+    telephoneEtudiant = request.form.get('telephoneEtudiant')
+    print(telephoneEtudiant)
+    composante = request.form.get('composante')
+    print(composante)
+    etablissement = request.form.get('etablissement')
+    print(etablissement)
+    #message = 'nom :' + nom + ' - groupe :' + groupe + ' pour l etudiant d identifiant :' + id
+    #print(message)
+    mycursor = get_db().cursor()
+    tuple_param = (nomEtudiant, prenomEtudiant, emailEtudiant, telephoneEtudiant, composante, idEtudiant)
+    sql = """UPDATE Etudiant SET nom=%s, prenom=%s, email=%s, telephone=%s, id_composante=%s WHERE id_etudiant=%s;"""
+    mycursor.execute(sql, tuple_param)
+    get_db().commit()
+    return redirect('/etudiant/show')
 #Contrat (Audrick)
 @app.route('/contrat/show')
 def show_contrats():
